+++ src/apps/toolbar/modules/main/ContextMenu.tsx
@@ -1,14 +1,15 @@
 import { IconName } from '@icons';
 import { invoke, SeelenCommand } from '@seelen-ui/lib';
 import { PluginId } from '@seelen-ui/lib/types';
-import { Flex, Menu, Popover } from 'antd';
+import { Checkbox, Flex, Menu, Popover } from 'antd';
+import { MenuItemType } from 'antd/es/menu/interface';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { BackgroundByLayersV2 } from 'src/apps/seelenweg/components/BackgroundByLayers/infra';
 
 import { RootActions, Selectors } from '../shared/store/app';
-import { SaveToolbarItems } from './application';
+import { RestoreToDefault, SaveToolbarItems } from './application';
 import { Icon } from 'src/apps/shared/components/Icon';
 
 export function MainContextMenu() {
@@ -41,14 +42,26 @@ export function MainContextMenu() {
                   <BackgroundByLayersV2 className="tb-context-menu-container">
                     <Menu
                       className="tb-context-menu"
-                      items={plugins
-                        .toSorted((p1, p2) => p1.id.localeCompare(p2.id))
-                        .map((plugin) => {
+                      items={[
+                        {
+                          key: 'restore',
+                          icon: <Icon iconName="TbRestore" />,
+                          label: t('context_menu.restore'),
+                          onClick() {
+                            RestoreToDefault();
+                          },
+                        },
+                        {
+                          type: 'divider',
+                        },
+                        ...plugins
+                          .toSorted((p1, p2) => p1.id.localeCompare(p2.id))
+                          .map<MenuItemType>((plugin) => {
                           const added = isAlreadyAdded(plugin.id);
                           return {
                             key: plugin.id,
                             icon: <Icon iconName={plugin.icon as IconName} />,
-                            label: added ? `✓ ${plugin.id}` : plugin.id,
+                            label: <Checkbox checked={added}>{plugin.id}</Checkbox>,
                             onClick: () => {
                               if (added) {
                                 dispatch(RootActions.removeItem(plugin.id));
@@ -58,13 +71,14 @@ export function MainContextMenu() {
                               SaveToolbarItems();
                             },
                           };
-                        })}
+                        }),
+                      ]}
                     />
                   </BackgroundByLayersV2>
                 }
               >
                 <Flex justify="space-between" align="center">
-                  {t('context_menu.add_module')}
+                  {t('context_menu.modules')}
                   <Icon iconName="FaChevronRight" size={12} />
                 </Flex>
               </Popover>